requirements.txt
# Core ML & data
numpy>=1.24
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
joblib>=1.3

# Dataset
datasets>=2.14
huggingface_hub>=0.20

# Vision / OCR
Pillow>=10.0
opencv-python>=4.8
pytesseract>=0.3.10
scikit-image>=0.21

# Deep learning (CNN)
torch>=2.0
torchvision>=0.15

# NLP — spaCy 3.8+ pulls thinc>=8.3.12, which requires Python 3.10+; 3.7.x works on 3.9
spacy>=3.7,<3.8

# PDF & API
pdfplumber>=0.10
flask>=3.0

# spaCy model (install separately):
# python -m spacy download en_core_web_sm
